feat: move quick register API from WalletAdminController to WalletManagementController
